feat: refactor LoyalitySupportProject component and add new task item components
