+++ src/components/Loyality/LoyalitySupportProject/LoyalitySupportProjectTaskItem/LoyalitySupportProjectTaskItem.module.scss
@@ -1,0 +1,101 @@
+@import "../../../../styles/mixins";
+@import "../../../../styles/vars";
+
+.loyalitySupportProjectTaskItem {
+  width: 100%;
+  padding: 1px;
+
+  background: linear-gradient(180deg, #6a4cd7 0%, #0d0623 100%);
+  transition: $transition;
+
+  &_completed {
+    background: $color-primary;
+    padding: 1px;
+  }
+
+  &__inner {
+    @include flexCol();
+    gap: 10px;
+    background: #090419;
+    padding: 7px;
+  }
+  &__main {
+    display: flex;
+    gap: 5px;
+    flex: 1;
+  }
+  &__img {
+    width: 20px;
+    height: 20px;
+    border-radius: 2px;
+    object-fit: cover;
+    border: 1px solid $color-primary;
+  }
+  &__texts {
+    display: flex;
+    flex-direction: column;
+    gap: 7px;
+    flex: 1;
+  }
+  &__name {
+    @include setFont(12px, 400);
+  }
+  &__description {
+    @include setFont(12px, 400, $font-fam-montserrat);
+    font-style: italic;
+  }
+  &__actions {
+    @include flexCol();
+    flex-direction: column;
+    gap: 6px;
+  }
+  &__subscribeBtn {
+    @include button();
+    @include setFont(12px, 700, $font-fam-main, $color-disabledText);
+    transition: $transition;
+    &:disabled {
+      color: #6a4cd780;
+    }
+  }
+  &__getBtn {
+    max-width: 100%;
+    width: 110px;
+    @include button($color-primary);
+    padding: 1px;
+    transition: $transition;
+
+    &:disabled {
+      padding: 0;
+    }
+  }
+
+  &__getBtnInner {
+    width: 100%;
+    @include flexCol();
+    padding: 8px;
+    gap: 1px;
+    @include setFont(8px, 400);
+    background: linear-gradient(180deg, #3d2b7e -35.56%, #040010 132.26%);
+    transition: $transition;
+  }
+
+  &__getBtn,
+  &__getBtnInner {
+    clip-path: polygon(0% 0%, 96% 0%, 100% 20%, 100% 80%, 96% 100%, 0% 100%);
+  }
+
+  &__getBtn:disabled &__getBtnInner {
+    background: linear-gradient(180deg, #3d2b7e -35.56%, #040010 132.26%);
+    color: #c6b7ff80;
+  }
+
+  &__getBtnImg {
+    width: 12px;
+    height: 12px;
+  }
+
+  &__adsgramWrapper {
+    background: #090419;
+    padding: 7px;
+  }
+}
